* Added attributes: format, noreturn * Fixed various bugs found with format attribute
svn: 981

+++ src/dcc.h
@@ -113,7 +113,7 @@ struct bot_info {
 
 struct relay_info {
   struct chat_info *chat;
-  int sock;
+  long sock;
   int old_status;
   port_t port;
 };
